Calculate number of ticks used for ballrun correctly (I hope). Changed ballrun message, now message will look like this: < The ball was loose for 155 frames / 155.00 ticks (best 67.00) / 12.92s >

+++ xpilot/src/server/rank.h
@@ -168,9 +168,9 @@ static inline void Rank_ballrun(player_t *pl, int tim)
 	pl->rank->entry.bestball = tim;
 }
 
-static inline int Rank_get_best_ballrun(player_t *pl)
+static inline double Rank_get_best_ballrun(player_t *pl)
 {
-    return (pl->rank ? pl->rank->entry.bestball : 65535);
+    return (double)(pl->rank ? pl->rank->entry.bestball : 65535);
 }
 
 static inline void Rank_add_ball_kill(player_t *pl)      { Rank_add_kill(pl); }
